pyproject.toml
[tool.poetry]
name = "django-semantic-admin"
version = "0.1.5"
description = "Django Semantic UI Admin theme"
authors = ["Alex <globophobe@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/globophobe/django-semantic-admin"
repository = "https://github.com/globophobe/django-semantic-admin"
keywords = ["django", "admin", "theme"]
classifiers = [
    "Framework :: Django",
    "Operating System :: OS Independent",
]
packages = [{ include = "semantic_admin" }]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
django = ">=3.0"
django-taggit = "^1.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.7.0"
Pillow = "^8.1.0"
django-stubs = "^1.7.0"
mypy = "^0.790"
django-filter = "^2.4.0"
invoke = "^1.5.0"
factory-boy = "^3.2.0"
Faker = "^5.3.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
